+++ server/routes/index.route.js
@@ -4,11 +4,9 @@ const instagramRoute = require("./instagram.route");
 const twitterRoute = require("./twitter.route");
 const facebookRoute = require("./facebook.route");
 const youtubeRoute = require("./youtube.route");
-const ResocieObs = require("../../config/resocie.json").observatory;
-const httpStatus = require("../../config/resocie.json").httpStatus;
+const portalRoute = require("./portal.route");
 
 const router = express.Router();
-
 // router.use("/", spreadsheetsRoute);
 
 router.get("/", (req, res) => {
@@ -54,52 +52,14 @@ router.get("/", (req, res) => {
 	});
 });
 
-router.get("/qualquer", (req, res) => {
-	const data = {
-		texto: "testandi",
-		extra: "qualquer coisa",
-		statusCod: httpStatus.ERROR_SPLIT_ACTORS,
-	};
-	// res.render("qualquer");
-	console.log("Passou por aqui");
-	console.log(`Status enviado = ${httpStatus.ERROR_SPLIT_ACTORS}`);
-	res.send(data);
-});
+router.use("/portal", portalRoute);
 
-router.get("/espacoExploratorio", (req, res) => {
-	const socialMedia = ResocieObs.socialMidia;
-	const medias = [];
-
-	for (midia in socialMedia) {					// eslint-disable-line
-		const aux = socialMedia[midia];				// eslint-disable-line
-		const list = {
-			lower: aux,
-			upper: capitalize(aux),
-		};
-
-		medias.push(list);
-	}
-
-	res.render("plot", { medias });
-});
-
-router.get("/espacoExploratorio2", (req, res) => {
-	res.render("plot2");
-});
-
-// mount facebook routes at /facebook
 router.use("/facebook", facebookRoute);
 
-// mount twitter routes at /twitter
 router.use("/twitter", twitterRoute);
 
-// mount youtube routes at /youtube
 router.use("/youtube", youtubeRoute);
 
 router.use("/instagram", instagramRoute);
-
-const capitalize = (str) => {
-	return str.replace(/\b\w/g, l => l.toUpperCase()); // eslint-disable-line
-};
 
 module.exports = router;
