+++ views/flexibilidades.pug
@@ -1,0 +1,1 @@
+extends layout
