+++ views/mixins.pug
@@ -1,0 +1,48 @@
+// Mixin for all files
+mixin insertLink(URL, text)
+	a(href=URL, target='_blank')=text
+
+// mixin for the page "Quem Somos"
+mixin insertDeveloper(developer)
+	.col-md-7.order-md-2.align-self-center
+		h4.featurette-heading #{developer.name}
+		p #{developer.text}
+			| Possui um portifólio no 
+			+insertLink(developer.github, "GitHub")
+			| . 
+			+insertLink(developer.cv, "Saiba mais")
+	.col-md-5.order-md-1
+		img.featurette-image.img-fluid.mx-auto.rounded(src=developer.img, alt=developer.name)
+
+// mixin for the page "Espaço Exploratório"
+mixin digitalMedia
+	label(for="digitalMedia") Mídia digital:
+		select#digitalMedia.custom-select.d-block.w-100(required="")
+			option(value="") Rede...
+			each media in medias
+				option(value = media.lower)= media.upper
+		.invalid-feedback
+			| Por favor selecione uma mídia válida.
+
+mixin queries
+	label(for="queries") Aspecto de análise:
+		select#queries.custom-select.d-block.w-100(required="")
+			option(value="") Escolha...
+		.invalid-feedback
+			| Por favor selecione um aspecto válido.
+
+mixin categories
+	label(for="categories") Categorias:
+		select#categories.custom-select.d-block.w-100(required="")
+			option(value = "") Escolha...
+			option(value = "F/C") Frentes/Coletivos
+			option(value = "OSC") Organizações da Sociedade Civil
+			option(value = "CP") Candidatos à presidência
+		.invalid-feedback
+			| Por favor selecione um aspecto válido.
+
+mixin actors
+	#actors.form-check
+		input#defaultCheck.form-check-input(type='checkbox', value='')
+		label.form-check-label(for='defaultCheck')
+			| Atores
