+++ views/plot.pug
@@ -30,37 +30,3 @@ block content
 							input#clearActors.form-check-input(type="checkbox")
 							label.form-check-label(for="clearActors") Limpar atores
 				button#search.btn.btn-primary(type='button') Visualizar
-
-
-mixin digitalMedia
-	label(for="digitalMedia") Mídia digital:
-		select#digitalMedia.custom-select.d-block.w-100(required="")
-			option(value="") Rede...
-			each media in medias
-				option(value = media.lower)= media.upper
-		.invalid-feedback
-			| Por favor selecione uma mídia válida.
-
-mixin queries
-	label(for="queries") Aspecto de análise:
-		select#queries.custom-select.d-block.w-100(required="")
-			option(value="") Escolha...
-		.invalid-feedback
-			| Por favor selecione um aspecto válido.
-
-mixin categories
-	label(for="categories") Categorias:
-		select#categories.custom-select.d-block.w-100(required="")
-			option(value = "") Escolha...
-			option(value = "F/C") Frentes/Coletivos
-			option(value = "OSC") Organizações da Sociedade Civil
-			option(value = "CP") Candidatos à presidência
-		.invalid-feedback
-			| Por favor selecione um aspecto válido.
-
-mixin actors
-	#actors.form-check
-		input#defaultCheck.form-check-input(type='checkbox', value='')
-		label.form-check-label(for='defaultCheck')
-			| Atores
-
